Removed old code that isn't called in the codebase from the ReadingListService. Ready to move the Controller logic into the Service.
