+++ Kavita.API/Services/ReadingLists/IReadingListService.cs
@@ -2,8 +2,6 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Kavita.Common.Helpers;
 using Kavita.Models.DTOs.ReadingLists;
-using Kavita.Models.DTOs.ReadingLists.CBL;
-using Kavita.Models.DTOs.ReadingLists.CBL.V1;
 using Kavita.Models.Entities;
 using Kavita.Models.Entities.Enums;
 using Kavita.Models.Entities.ReadingLists;
@@ -31,8 +29,6 @@ public interface IReadingListService
     /// <param name="library"></param>
     /// <returns></returns>
     Task CreateReadingListsFromSeries(Series series, Library library);
-
-    Task CreateReadingListsFromSeries(int libraryId, int seriesId);
     Task<string> GenerateReadingListCoverImage(int readingListId);
     /// <summary>
     /// Check, and update if needed, all reading lists' AgeRating who contain the passed series
@@ -42,7 +38,5 @@ public interface IReadingListService
     /// <returns></returns>
     /// <remarks>This method does not commit changes</remarks>
     Task UpdateReadingListAgeRatingForSeries(int seriesId, AgeRating ageRating);
-
     Task<IList<ReadingListItemDto>> GetReadingListItems(int readingListId, int userId, UserParams? userParams = null);
-    Task<ReadingListItemDto?> GetContinueReadingPoint(int readingListId, int userId);
 }
